backend/requirements.txt
asgiref==3.8.1
cmake==4.0.0
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
Markdown==3.7
numpy==2.2.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
sqlparse==0.5.3
tzdata==2025.2
